Added a floppy disk icon
Just noticed that one of the lines is slightly skewed. Fml.

I was going to upload a much higher quality variant of this image, but
then I didn't, because there is no need. I wouldn't want users to have
to download a 20 megabyte file for a couple home screen icons (not to
mention the loading time). Also the image probably would've increased
the repository size by like 5 fold (if the note icon in the app hasn't
already).

Later, users will be able to select whether they want the higher quality
images or not. (Just in case they have a 32k screen and would like to
render an 8k floppy disk in the span of 1.5x1.5 physical in^2.

Why did I make this?...